Fix issue #1333
The Projectsend release r1720 introduced the folder implementation. However, this feature is not very useful if clients cannot see the folders and their contents!

We use Projectsend to share numerous documents related to our products with customers. Until now, we have only used categories and category filters to reduce the file lists. With the new folder feature, the structure has become even better!

The folders in Projectsend are not traditional directories but virtual folders designed to improve organization for admins and clients. All files remain in the upload directory, and the folder information is stored in the `tbl_folders` table.

**Problem:** When I upload new files as an admin or operator and assign these files to clients and client groups, the clients cannot see the files in the designated folders. They can only see files in the root folder.

**Analysis:** In the `getFolders` function within the `Folders` class, the `public_or_client` argument is not set. As a result, folders that contain files assigned to specific clients or client groups are not displayed.

**Solution:**

1. Check the access level (`level`): if `level = 0`, it indicates a client; otherwise, it’s an admin or operator.
2. If `level ≠ 0`, keep the behavior as it is.
3. If `level = 0`, retrieve the client ID. Display all folders that contain files assigned to the client or to a group of which the client is a member.
4. Set the folders as “public” so that clients can see the contents. The permissions remain attached to the files themselves, not to the folders. (I am not sure how the future folder functionality is planned, but this setup is perfect for our needs. I hope this will be available as an option in a future release.)
5. Empty folders or folders without assigned files will be hidden from the client. If a client has a file in a sub-folder, all parent directories up to the root will be rendered and shown.